requirements.txt
# Por medio del comando pip freeze se recolpilo la siguiente informacion de lls requerimientos actuales para el proyecto actual

asgiref==3.5.0
Django==4.0.3
djangorestframework==3.13.1
pytz==2022.1
sqlparse==0.4.2
tzdata==2022.1